backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-chess==1.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
